Backend/pyproject.toml
[tool.poetry]
name = "fastapi_demo"
version = "0.1.0"
description = ""
authors = ["hong9201@gmail.com"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.72.0"
uvicorn = "0.17.4"
pydantic = "^1.3"
requests = "^2.22.0"
loguru = "^0.4.0"
joblib = "^0.15.1"
python-multipart = "^0.0.5"
Jinja2 = "^3.1.2"
google-cloud-storage = "^2.4.0"
pandas = "^1.4.2"
tqdm = "^4.64.0"
torch = "^1.11.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = {version = "^19.10b0", allow-prereleases = true}
autopep8 = "^1.4.4"
ipdb = "^0.12.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
